rename plugin to ua-restriction
